[1890][eval] Move MergeReader to own module (#1892)
* move mergereader

* use assertions only
