+++ packages/evaluate/src/weathergen/evaluate/run_evaluation.py
@@ -28,9 +28,9 @@ from weathergen.common.config import _REPO_ROOT
 from weathergen.common.logger import init_loggers
 from weathergen.common.platform_env import get_platform_env
 from weathergen.evaluate.io.csv_reader import CsvReader
+from weathergen.evaluate.io.merge_reader import WeatherGenMergeReader
 from weathergen.evaluate.io.wegen_reader import (
     WeatherGenJSONReader,
-    WeatherGenMergeReader,
     WeatherGenReader,
     WeatherGenZarrReader,
 )
